+++ Protocol/GlkEventListenerProtocol.h
@@ -17,7 +17,7 @@
 @protocol GlkEventListener <NSObject>
 
 /// Called by the session object whenever an event arrives
-- (oneway void) eventReady: (int) syncCount;
+- (oneway void) eventReady: (NSInteger) syncCount;
 
 @end
